refactor(Test) : 테스트 코드 메서드 추가 및 공백 정리, DirtiesContext 추가
